Rename src/lolex.js to src/lolex-src.js to avoid bithound ignoring it
