+++ src/lolex-src.js
